use int instead of int6 to reduce int promotions overhead if any during bit ops

+++ src/common.h
@@ -58,8 +58,8 @@ typedef struct unz__bitstate_t {
   struct unz__chunk_t *chunk;
   bitstream_t          pbits; /* back buff  */
   BITS_TYPE            bits;  /* front buff */
-  uint_fast16_t        nbits;
-  uint_fast16_t        npbits;
+  int                  nbits;
+  int                  npbits;
 } unz__bitstate_t;
 
 struct unz__stream_t {
